requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
pydantic>=2.8.0
python-multipart>=0.0.9
jinja2>=3.1.4
python-docx>=1.1.2
httpx>=0.27.0
pytest>=8.0.0
openai>=1.40.0
python-dotenv>=1.0.1
PyYAML>=6.0.1
playwright>=1.44.0
beautifulsoup4>=4.12.3
